# lib/api-zod/src/generated/types/cartItemInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Soko Kenya e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface CartItemInput {
  sessionId: string;
  productId: number;
  /** @minimum 1 */
  quantity: number;
}