# lib/api-zod/src/generated/types/cartItemUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Soko Kenya e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface CartItemUpdate {
  /** @minimum 0 */
  quantity: number;
}