# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Soko Kenya e-commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryOrdersByStatusItem } from './dashboardSummaryOrdersByStatusItem';

export interface DashboardSummary {
  totalOrders: number;
  totalRevenueKes: number;
  totalProducts: number;
  totalCategories: number;
  pendingOrders: number;
  ordersByStatus: DashboardSummaryOrdersByStatusItem[];
}