# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Soko Kenya e-commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';

export interface Order {
  id: number;
  sessionId: string;
  status: string;
  totalKes: number;
  phone: string;
  county: string;
  town: string;
  address: string;
  /** @nullable */
  customerName?: string | null;
  items: OrderItem[];
  createdAt: string;
}