# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Soko Kenya e-commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductListingStatus } from './productListingStatus';
import type { ProductOwnerType } from './productOwnerType';

export interface Product {
  id: number;
  name: string;
  description: string;
  priceKes: number;
  category: string;
  imageUrl: string;
  stock: number;
  /** @nullable */
  sellerName?: string | null;
  /** @nullable */
  county?: string | null;
  rating: number;
  reviewCount: number;
  featured: boolean;
  ownerType: ProductOwnerType;
  /** @nullable */
  merchantId: number | null;
  listingStatus: ProductListingStatus;
  isSponsored: boolean;
  /** @nullable */
  sponsoredUntil: string | null;
  /** @nullable */
  promotionTier: string | null;
  createdAt: string;
}