# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Soko Kenya e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  description: string;
  /** @minimum 0 */
  priceKes: number;
  category: string;
  imageUrl: string;
  /** @minimum 0 */
  stock: number;
  sellerName?: string;
  county?: string;
  featured?: boolean;
}